Ref: Events are now dealt with FXML
This whole refactoring has for purpose to switch the way events are
handle. Since you can use the @FXML annotation to get your components
back and their respective events, it is now much cleaner. Each event are
now dealt in separate methods, instead of all of them being called in
the initialize() method.
Also cleaned the FXML so the whole CSS is actually in the style.css
file.
Removed all commented and WIP code as i wanted a clean copy of the
project.