Ftr:Add the ability to change the state of a ToDo.
This is probably the last UI feature. Adding this feature needed to use the CellFactory method to let us build
the cell the way we needed it, meaning with a button inside the cell. This change also helped us keep the
ListView componenet as our display for the ToDos.
Overriding the updateItem method from the ListCell object was necessary for that, but it got rather clean with
the code managed the way it is.